+++ src/app/components/portfolio-artwork/portfolio-artwork.component.ts
@@ -73,7 +73,7 @@ export class PortfolioArtworkComponent implements OnInit, OnDestroy {
       console.log("Selected Art: " + JSON.stringify(currArt))
       this.artwork = currArt
       //this.fileName = this.artwork.images.filename
-      this.imageSRC = this.artwork.images.imageBase64
+      //this.imageSRC = this.artwork.images.imageBase64
       this.prev_image = this.artwork.images.imageBase64
       this.initForm()
     })
@@ -209,7 +209,7 @@ export class PortfolioArtworkComponent implements OnInit, OnDestroy {
       this.portfolioForm.reset();
       //console.log("prev_image " + JSON.stringify(this.prev_image))
       //console.log("imageSRC " + JSON.stringify(this.imageSRC))
-      if(this.imageSRC !== this.prev_image){
+      if(this.imageSRC !== this.prev_image && this.imageSRC){
         this.afStorage.storage.refFromURL(this.imageSRC).delete();
       }
       this.imageSRC = '';
@@ -263,7 +263,7 @@ export class PortfolioArtworkComponent implements OnInit, OnDestroy {
       if (userdata) {
         console.log("On Save Art: " + JSON.stringify(this.portfolioForm))
         //this.portfolioForm.get('artowkimage')?.reset();
-        if(this.imageSRC != this.prev_image ){
+        if(this.imageSRC != this.prev_image && this.prev_image ){
           this.afStorage.storage.refFromURL(this.prev_image).delete();
         }
         //this.afStorage.storage.refFromURL(this.prev_image).delete();
